Create 2 children in the header and start styling
The 2 children will contain the major parts of the
header - the top and bottom. Also Create 2 children
for the top div - one to contain the search icon
with the search input and another to contain
the profile elements. Also add the search icon png.

Give the header-top a display of grid and 2 columns
of 2fr and 1fr respectively because the search div
should be wider than the profile div.

Align the search icon with align-self: center
so it lines up properly with the search input.

Style the search input with a color, a
border-radius and a padding.