Add the items on the profile div and style
Add the bell icon for the notification,
the image avatar for the profile picture and
the profile name.

Display the profile div with display: grid.
Divide it into 3 columns of 1fr each to contain
the 3 items in the profile div.

Style the bell icon with a fixed height and
justify horizontally to the end of its track
so it gets closer to the avatar. Also align it
vertically center so it lines up properly with
the avatar.

Style the avatar with a border radius to make
it completely round and a fixed height. Also
justify it horizontally to the center so it aligns
with the notification icon and profile name

Align the profile name with align-self: center
to line it up with the rest of items in the
profile div.

A bigger value for the height of the avatar makes
the container wider and displaces the search icon
from lining up with the search input. Justify it
to the end of its track so it gets closer to the
search input.