Add the header-bottom items and style
Create two direct descendants for the header-bottom.
The first contains the larger avatar and the welcome
text. The second will contains header buttons.

Style the header-bottom with display: grid and
divide it into two columns to contain the
.welcome-message and the .header-buttons.

Also style the .welcome-message with a grid
display and divide it into two columns of
0.25fr and 1fr respectively. The 0.25fr is so
that the space will just be enough for the
avatar and so that any space between the avatar
and the welcome-text won't be much. Remove the
justify-self: end from the so that the avatar
search icon so that it can align with the avatar
along the left edge.

Give the bigger avatar a fixed height and a
border-radius

Give the .header-buttons a display of grid
and divide it into 3 columns of 1fr each
to contain the three buttons. Align the
buttons to the center so they can align with
the items of the .welcome-message. Also give
the .header-buttons a gap so there can be
an equal amount of space between the buttons.

Give the buttons themselves a fixed height,
a border-radius and a remove their borders.