Re-align the search icon and search input
Reduce the padding of the search input to make
the input smaller and align it to the center.
Justify the search icon to the center to
align it with the search input. This also ensures a
bit of space between it and the search input.

Style the .header with a display of grid and
divide it into rows so that the space between
.header-top and header-bottom can be
controlled.