Reduce the items and space in the .header
The items and space in the .header were much
and caused much vertical overflow to the rest
of the content.

To reduce the space consumed by the .header,
reduce the padding top and bottom, reduce the
heights of the notification icon and avatar.
Also reduce the height of the .welcome-message
avatar.

With the height of the .welcome-message avatar
reduced, also reduce the column size of the
column track containing it so that it can
occupy as much space as it needs and remain
close to the welcome text.

Also align the .profile avatar to the center to
the center to align it properly with the both
items on its right and the left.