Add the .sidebar items and style
Create three direct children for the
sidebar. THe 3 children contain the
.dashboard-header, .dashboard-main,
.dashboard-last.

Style the dashboard header as a grid
and divide it into 2 columns for the
icon and the text. The text column is
wider because the text is wider

Style the dashboard icon with a fixed
height and justify it to the end of
its column track so that it gets
closer to the dashboard text.

Also remove the padding left and right
from the .header and use a
justify-content: space-around to create
space on the sides of the header.