Add items to .sidebar-main and style
First rename the classes of 3 direct children
of .sidebar and updated the names in the CSS.

Add icons and text to the sidebar-main. Give
the .sidebar-main a grid display, a gap and
divided it into 2 columns - one for the icons
and the other for the paragraph. Style the
icons with a fixed height and
justify-self: end to bring them a bit closer to
the paragraphs