Reduce .sidebar-main gap and align items
Reduce the gap between .sidebar-main's
items so that the icons can be a bit closer
to the text. Also align sidebar-main items
to center so that the paragraphs can line up
with the icons.

Relocate the declaration of the rules of
the parent .sidebar to higher up in the CSS
so that the parent is the first mention of
sidebar.