Add the .main items and style
Create two direct children of .main.
.main-left contains the
.main-left-title and
the .content. The .content contains the cards.
the .main-right will contain the Announcements
and Trending sections.

Style the .main with a display of grid and
divide it into 2 columns of 2fr and 1fr
to contain the bigger .main-left and the smaller
.main-right.

Style the .main-left with a display of grid and
divide it into two rows so it can contain the
.main-left-title and the .content as blocks.