Style the .main-left-title and .content
Self align the .main-left-title to the center
so it can maintain a bit of space between it and
the .content.

Give the content a grid display and use the
repeat function to auto-fit the cards, using
the constraints of the minmax function. The
cards will assume a minimum width of 250px
and if there is more space, they would divide
the space equally - 1fr each, of their container.