Add the .trending content and style
Create the .trending-title and .trending-content
as direct children of the .trending section.
The .trending-content contains the avatars
and their respective handles.

Group the style declarations of .announcements
and .trending together because they have the
same styles.

Style the avatar images with a fixed height and
a border radius to make them round.