Fill the .sidebar svgs with color
Give the .sidebar SVGs inline style
of fill: white.

Could place the SVGs inline in the HTML
document which makes it more customizable
can be styled with a class selector.
This was not done because the fill property
is the only needed customization and need
a clear, easier-to-read HTML.

Add border-left styles for the article cards.