Style the announcement and trending sections
Create CSS variable for the two shades of
colors and ruse it to replace the colors
in .main and .sidebar. Also style the header
buttons.

In the HTML, wrap the article texts below
the title in divs so that they can be targeted
and styled with a different color from the
titles.