Add box shadow to the header.
Give the .header a position of relative
before applying box-shadow. Any
box-shadow applied will not be visible
as until .header is positioned because
.main renders immediately below header.

Reduce the .welcome-message image height
to create more space in the document.
Adjust the column size of .welcome-message
so that the image's column is narrow.
This makes the .welcome text to get a bit
closer to the image.