Reduce size of elements and add padding
This is to make these elements smaller
so that the page does not overflow its
current viewport height. Reduce the
font size of the .sidebar-header
paragraph and reduce the heights of
the sidebar images. Also reduce the
font-sizes of the announcement headers
headers and announcement text.

Add bottom padding to the .main .
Also Add left and right padding
around the announcement card content
and at the bottom of the announcement text