Properly align items in .trending
Reduce the sizes of the trending content
avatars and adjust the size of the
columns of the avatars in the .trending-content.
This would bring them a little closer to
the .avatar-content. Give the .trending-content
a padding-left to put some space between the
avatars and the left edge of the container

Align the avatars and the avatar handles to the
center to make them align horizontally.