Change the track size of header-buttons column
The buttons were too wide because their column
shared equal track size with the welcome message.
The buttons then stretched out and were too wide.
Reduce the size of the column for the buttons
on the .header-bottom. In .header-buttons,
use the repeat function to create 3 buttons
with fixed width - some lower value. Buttons
are now less wide

The adjustment of column size in the
.header-bottom has increases the space between
the welcome message avatar and the welcome
text because the .welcome-message has more
space to spread out wide.
Justify-self: center on the avatar brings it closer
to the welcome text.

Align the .sidebar-header paragraph to the center
so that it can line up better horizontally with
the sidebar header icon.

Give the article .card-icons a padding top
to push them a but away from the bottom
of the article.

Correctly spell the complementary color variable
name and apply the correction as needed.
Change the the complementary color.